+++ app/src/main/java/co/develhope/meteoapp/features/home/Home.kt
@@ -1,23 +1,25 @@
-package co.develhope.meteoapp
+package co.develhope.meteoapp.features.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import co.develhope.meteoapp.databinding.FragmentTodayBinding
-import co.develhope.meteoapp.databinding.FragmentTomorrowBinding
+import co.develhope.meteoapp.R
+import co.develhope.meteoapp.databinding.FragmentHomeBinding
 
+class Home : Fragment() {
 
-class Tomorrow : Fragment() {
-    private lateinit var binding: FragmentTomorrowBinding
+    private lateinit var binding: FragmentHomeBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentTomorrowBinding.inflate(inflater)
+
+        binding = FragmentHomeBinding.inflate(inflater)
         return binding.root
     }
+
 }
