+++ app/src/main/java/co/develhope/meteoapp/features/today/Today.kt
@@ -1,4 +1,4 @@
-package co.develhope.meteoapp
+package co.develhope.meteoapp.features.today
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,10 +6,11 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import co.develhope.meteoapp.databinding.FragmentHomeBinding
-import co.develhope.meteoapp.databinding.FragmentSearchBinding
+import co.develhope.meteoapp.databinding.FragmentTodayBinding
 
-class Search : Fragment() {
-    private lateinit var binding: FragmentSearchBinding
+class Today : Fragment() {
+
+    private lateinit var binding: FragmentTodayBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -17,7 +18,7 @@ class Search : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
 
-        binding = FragmentSearchBinding.inflate(inflater)
+        binding = FragmentTodayBinding.inflate(inflater)
         return binding.root
     }
