+++ app/src/main/java/co/develhope/meteoapp/features/tomorrow/Tomorrow.kt
@@ -1,25 +1,23 @@
-package co.develhope.meteoapp
+package co.develhope.meteoapp.features.tomorrow
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import co.develhope.meteoapp.R
-import co.develhope.meteoapp.databinding.FragmentHomeBinding
+import co.develhope.meteoapp.databinding.FragmentTodayBinding
+import co.develhope.meteoapp.databinding.FragmentTomorrowBinding
 
-class Home : Fragment() {
 
-    private lateinit var binding: FragmentHomeBinding
+class Tomorrow : Fragment() {
+    private lateinit var binding: FragmentTomorrowBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-
-        binding = FragmentHomeBinding.inflate(inflater)
+        binding = FragmentTomorrowBinding.inflate(inflater)
         return binding.root
     }
-
 }
