+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/blue"
-    tools:context=".Home">
+    tools:context=".features.home.Home">
 
     <TextView
         android:layout_width="wrap_content"
