+++ app/src/main/res/layout/fragment_search.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/teal_700"
-    tools:context=".Search">
+    tools:context=".features.search.Search">
 
     <TextView
         android:layout_width="wrap_content"
