+++ app/src/main/res/layout/fragment_today.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/teal_200"
-    tools:context=".Today">
+    tools:context=".features.today.Today">
 
     <TextView
         android:layout_width="wrap_content"
