+++ app/src/main/res/layout/fragment_tomorrow.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/purple_200"
-    tools:context=".Tomorrow">
+    tools:context=".features.tomorrow.Tomorrow">
 
     <TextView
         android:layout_width="wrap_content"
