+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,25 +7,25 @@
 
     <fragment
         android:id="@+id/fragment_home"
-        android:name="co.develhope.meteoapp.Home"
+        android:name="co.develhope.meteoapp.features.home.Home"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/fragment_today"
-        android:name="co.develhope.meteoapp.Today"
+        android:name="co.develhope.meteoapp.features.today.Today"
         android:label="fragment_today"
         tools:layout="@layout/fragment_today" />
 
     <fragment
         android:id="@+id/fragment_tomorrow"
-        android:name="co.develhope.meteoapp.Tomorrow"
+        android:name="co.develhope.meteoapp.features.tomorrow.Tomorrow"
         android:label="fragment_tomorrow"
         tools:layout="@layout/fragment_tomorrow" />
 
     <fragment
         android:id="@+id/fragment_search"
-        android:name="co.develhope.meteoapp.Search"
+        android:name="co.develhope.meteoapp.features.search.Search"
         android:label="fragment_search"
         tools:layout="@layout/fragment_search" />
